feat: script to generate transparent capital colors
Automated way to generate different colorized capitals of existing
capitals. This as an simpler and more automated way than creating
new colored versions in GIMP. ImageMagik is used to colorize.

The interface of the script is designed in a way that it can be
integrated with other scripting in the future.

Documentation for the directory is also added.

Signed-off-by: Nico Rikken <nico.rikken@fsfe.org>