chore: add version number to Readings_Organisation.md
Introduce a version number to check on translations. This file is the most
relevant to be translated, as it can help volunteers in organising a reading.

The version number indicates whether translations are in sync with the main
text.

For the future it might be an option to transfer the file to the website to
benefit from those translation efforts.

Signed-off-by: Nico Rikken <nico.rikken@fsfe.org>