fix!: revert templates 23rd of May
Revert the template changes done in commit 3423c3bf26, which were done to accomodate the longer sentences of the new Portugese translation. The template change for the print version however positioned the first text block of page 8 too low on the page, breaking rendering for all translations. This commit revers the templates to a known good state, from which a solution can be sought to accomodate the Portugese translation. This can be a change in the source text, in the template, or both. Regardless the template change in the print and screen templates should be similar to prevent drift between the templates.

This commit is related to issue #24 about the alignment on page 8.

Thanks to Petter Reinholdtsen for pointing out the issue.

Link to previous changeset in FSFE git:
https://git.fsfe.org/FSFE/ada-zangemann/commit/3423c3bf26523301ef810b214bfb82145ab74531#diff-c28a82ee49a9f471bf6cc06b342c0384d9d4b546

Signed-off-by: Nico Rikken <nico.rikken@fsfe.org>