feat: headers using fonts
Updated print template to write titles using fonts instead of images.

Additional text frames are added, to align with images.

Fonts used: Amatic-SC and Delicious Handrawn, depending on the original writing style.

The template was updated in Scribus 1.6.2, which results in errors when trying to view or render in previous Scribus versions.

A common image was created for page 3 that lacks a subtitle. The image could be optimised reduced in size.

Signed-off-by: Nico Rikken <nico.rikken@fsfe.org>