chore: remove page-level section
Remove section for page. Simplesect used for 2 text frames on page 8. As
simplesect elements cannot be nesed a section was required to model a page. This
has the downside of ending up in the final output, which is undesireable. As
there is no current need to model a section, it can be removed.

Signed-off-by: Nico Rikken <nico.rikken@fsfe.org>