feat: no text on ice cream page
The ice cream page contains a text about the website. The Scribus template
removes this text by placing a white box over the text. This change remove the
text. It is kept for the German translation in case it is needed. If not it can
be removed in the future.

Signed-off-by: Nico Rikken <nico.rikken@fsfe.org>