feat: Scribus print file with inside cover
Print file with four pages added in the front and four in the back to include
the inside covers. The to_scribus.sh Perl script resolves capitals based on
page numbers, so rendering the template doesn't fully work. The resulting print
template was created by adding 4 pages in the front and back and add an image
frame that is adjusted to match the image and aligned in the center of the
spread.

Signed-off-by: Nico Rikken <nico.rikken@fsfe.org>