feat(templates/coverhard): enable font-based titles
Enable image and text outputs of coverhard template. Similar to the
template-print.sla setup. A single condition is used that combines print/web
and text/img. This is the easiest way to deal with combinations of conditions.
Recommendation for more advanced logic is to use multiple profiling steps, as
per https://sagehill.net/docbookxsl/MultiProfile.html That might be an option
for the future if that makes sense.

Idea for improvement is to manage Scribus files from a central place.

Signed-off-by: Nico Rikken <nico.rikken@fsfe.org>