feat(docbook): simplified colophon
Simplify the colophon (imprint) based on the English screen pdf.

This in order to address the issue of too long colophon pages becoming
partially covered by the grey box
https://git.fsfe.org/FSFE/ada-zangemann/issues/65

This removes details. The English pdf contains a revision, but this is ommitted
as it the source text is static.

The text in the grey box is adjusted as well.

Signed-off-by: Nico Rikken <nico.rikken@fsfe.org>