feat(docbook): remove unused data
The initial proposed docbook design included information about fonts and
credits. At the moment this information is not really used, and only confuses
translators. The credits might show up in metadata of Docbook output, that is
the only use-case.

By removing this information for the moment, the docbook file becomes simpler
and complexity for translators is reduced.

This information might return in the future if there is a need.

This information could also be left to an editor that adds this information
after the translation is done, much like the approach suggested in
https://git.fsfe.org/FSFE/ada-zangemann/issues/138

Signed-off-by: Nico Rikken <nico.rikken@fsfe.org>