feat(docbook/templates): support two colophon variants in print template
Adjust the template to support a colophon-print variant of the template which
is the full size of the page. The text field is a bit longer then the standard
page margins. Makefile targets are added to match. The simple colophon is kept
as default and is used for the existing target filenames.

Signed-off-by: Nico Rikken <nico.rikken@fsfe.org>