feat(docbook/makefile): fail when messages cannot be merged
In normal behavior itstool only warns if a message cannot be merged due to the
text not being XML and then falls back to using the original English text. This
behavior is undesirable because it means that a typo can cause the merge to
fail silently, resulting in some translation missing in the output.

A typical error is to not write the ampersand & as &amp; as the ampersand is a
special escape character in XML.

Using the strict behavior of itstool the message merge will fail and will
higlight the specific character where it went wrong.

This will not change the behavior of missing translations, where the merge will
fall back to the original English text.

Fixes issue https://git.fsfe.org/FSFE/ada-zangemann/issues/207

Signed-off-by: Nico Rikken <nico.rikken@fsfe.org>